requirements.txt
transformers==4.52.4
pytest==8.4.0
python-dotenv==1.1.0